update
update

+++ go-here/backend/server.js
@@ -348,12 +348,18 @@ app.get("/getAllRequested", express.json(), async (req, res) => {
 
 app.post("/postWashroomRequest", express.json(), async (req, res) => {
   try {
-    const { title, address, longitude, latitude } = req.body;
+    const { title, address, longitude, latitude, requestType } = req.body;
 
-    if (!title || !address || !longitude || !latitude) {
+    if (!title || !address || !longitude || !latitude || !requestType) {
       return res
         .status(400)
-        .json({ error: "Title, Address, Longitude, and Latitude are required." });
+        .json({ error: "Title, Address, Longitude, and Latitude and Request Type are required." });
+    }
+
+    if (!['', '', ''].includes(requestType.toUpperCase())) {
+      return res
+        .status(400)
+        .json({ error: "Invalid request type. It must be USERREQUEST, USERCLOSURE, or BUSINESSREQUEST." });
     }
 
     const collection = db.collection(COLLECTIONS.requested);
@@ -361,7 +367,8 @@ app.post("/postWashroomRequest", express.json(), async (req, res) => {
       title,
       address,
       longitude,
-      latitude
+      latitude,
+      requestType: requestType.toUpperCase()
     });
     res.json({
       response: "Washroom request added succesfully.",
@@ -392,28 +399,38 @@ app.patch("/patchRequestStatus/:washroomId", express.json(), async (req, res) =>
 
     if(status === "ACCEPTED"){
       const washroomData = await requestCollection.findOne(new ObjectId(washroomId));
-      const newWashroomData = await washroomCollection.insertOne({
-        title: washroomData.title,
-        address: washroomData.address,
-        longitude: washroomData.longitude,
-        latitude: washroomData.latitude,
-        sponsorlvl: 0
-      });
-      res.json({
-        response: "Washroom added succesfully.",
-        insertedId: newWashroomData.insertedId,
-      }); 
-      const requestData = await requestCollection.deleteOne({
-        _id: new ObjectId(washroomId),
-      });
 
-      if (newWashroomData.matchedCount === 0 || washroomData.matchedCount === 0) {
+      if (!washroomData) {
+        return res.status(404).json({error: "Requested washroom not found."});
+      }
+
+      if (washroomData.requestType === "USERCLOSURE") {
+        const deletedWashroom = await washroomCollection.deleteOne({ _id: new ObjectId(washroomId) })
+      } else {
+        const newWashroomData = await washroomCollection.insertOne({
+          title: washroomData.title,
+          address: washroomData.address,
+          longitude: washroomData.longitude,
+          latitude: washroomData.latitude,
+          sponsorlvl: 0,
+          requestType: washroomData.requestType
+        });
+        res.json({
+          response: "Washroom added succesfully.",
+          insertedId: newWashroomData.insertedId,
+        }); 
+        const requestData = await requestCollection.deleteOne({
+          _id: new ObjectId(washroomId),
+        });
+  
+        if (newWashroomData.matchedCount === 0 || washroomData.matchedCount === 0) {
+          return res
+            .status(404)
+            .json({ error: "Unable to find accepted washroom with given ID." });
+        } 
+        
         return res
-          .status(404)
-          .json({ error: "Unable to find accepted washroom with given ID." });
-      } 
-      
-      return res
+      }
     } else if (status === "DECLINED"){
       const requestData = await requestCollection.deleteOne({
         _id: new ObjectId(washroomId),
